se cambia tipo de dato que guarda fecha

+++ src/main/java/com/blo/sales/business/dto/DtoIntDebtor.java
@@ -17,7 +17,7 @@ public @Data class DtoIntDebtor implements Serializable {
 	
 	private BigDecimal total; //deuda que tiene
 	
-	private int open_date;
+	private long open_date;
 	
 	private List<BigDecimal> partial_pyments;
