+++ src/main/java/com/blo/sales/facade/dto/DtoDebtor.java
@@ -16,7 +16,7 @@ public @Data class DtoDebtor implements Serializable {
 
 	private BigDecimal total; // deuda que tiene
 
-	private int open_date;
+	private long open_date;
 
 	private List<BigDecimal> partial_pyments;
